refactor settings reading
.. in order to add LoadOneSetting for individual commandline settings
overrides.

This makes it necessary to also factor out LoadSettingsPost, in order
to do other work necessary after changing some specific parameters.

newMandelboxParametersLoaded was made into a global, though instead it
should probably be placed in sParamRender.  However, this is not too grave
a compromise as paletteLoadedFromSettingsFile is another global that could
be in the structure instead.